+++ multimux/src/lib.rs
@@ -32,6 +32,7 @@ mod http;
 pub mod origin;
 pub mod output;
 pub mod prometheus;
+pub mod push;
 mod redact;
 pub mod registry;
 pub mod route;